+++ src/pages/Signup/SignupForm/index.ts
@@ -1,0 +1,5 @@
+export { default as SignupForm } from './SignupForm';
+export { default as SignupUserName } from './SignupUserName';
+export { default as SignupPassword } from './SignupPassword';
+export { default as SignupPasswordConfirm } from './SignupPasswordConfirm';
+export { default as SignupTel } from './SignupTel';
